requirements.txt
# Django stuff
Django>=4.0
semver==2.6.0
pytz==2017.2
# PIL support
Pillow==8.2.0
python-dateutil==2.7.5

# mysql database
mysqlclient==1.4.6

django-storages==1.9.1
python-resize-image==1.1.19

# CSky models
django-cachemodel==2.2.0
django-basic-models==4.0.0

django-object-actions==3.1.0

python-memcached==1.58

djangorestframework==3.12.2

# Django Allauth
django-allauth
oauthlib==3.1.0
requests==2.25.0
requests-oauthlib==0.4.2
# Permissions framework
rules==0.4

# OAuth
django-oauth-toolkit==2.0.0

# Some extras
django-cors-headers==3.3.0
django-autoslug==1.9.8
puremagic==1.6
sqlparse==0.3.0
netaddr
defusedxml

# Utilities for working with badges
jsonschema==2.6.0
simplejson==3.6.4

# JSON-LD
PyLD==0.7.1
rfc3987
pypng==0.0.20
jsonfield==3.1.0

# markdown support
Markdown==2.6.8
django-markdownify==0.1.0
bleach==3.3.0

# Open Badges related utilities
badgecheck
# TODO replace badgecheck with openbadges after openbadges next release resolving dependency conflicts
# openbadges==1.1.1
openbadges-bakery

celery==4.4.0
django-celery-results==2.0.0
python-json-logger==0.1.2

# SSL Support
cffi==1.14.5
cryptography==3.3.2
idna==2.8
ipaddress==1.0.14
pyasn1==0.1.9
pycparser==2.14
six

#svg 2 png support
CairoSVG==1.0.22
cairocffi==0.9.0
cssselect==0.9.2
tinycss==0.4

# Backport of Python 3.0 CSV package with Unicode support
backports.csv==1.0.4

apispec==6.0.0
# apispec-djangorestframework==1.1.1

responses==0.12.1
mock==3.0.5
django_mock_queries==2.1.5

# lti consumer
lti==0.9.4

# Optional AWS dependencies
boto==2.49.0
boto3==1.14.35
django-boto==0.3.12

pysaml2==6.5.1
xmlsec==1.3.3

#locking a dependency lib
requests-cache==0.5.2
pymemcache
geopy==2.2.0
